🌀 added anchor investor registration

+++ bfn-atlas2/clients/src/main/kotlin/com/bfn/client/dto/InvestorProfileStateDTO.kt
@@ -1,12 +1,11 @@
 package com.bfn.client.dto
 
-import java.math.BigDecimal
 import java.util.*
 
 data class InvestorProfileStateDTO (val issuedBy: String?,
                                     val accountId: String,
-                                    val minimumInvoiceAmount: BigDecimal,
-                                    val maximumInvoiceAmount: BigDecimal,
-                                    val totalInvestment: BigDecimal,
+                                    val minimumInvoiceAmount: Double,
+                                    val maximumInvoiceAmount: Double,
+                                    val totalInvestment: Double,
                                     val defaultDiscount: Double,
                                     var date: Date)
