+++ bfn-atlas2/clients/src/main/kotlin/com/bfn/client/local/Client.kt
@@ -16,7 +16,6 @@ import net.corda.core.node.services.vault.QueryCriteria
 import net.corda.core.utilities.NetworkHostAndPort
 import net.corda.core.utilities.loggerFor
 import org.json.JSONArray
-import java.math.BigDecimal
 import java.util.*
 import kotlin.collections.MutableList
 import kotlin.collections.MutableMap
@@ -565,9 +564,9 @@ private class Client {
                 issuedBy = "thisNode", accountId = it.state.data.identifier.id.toString(),
                 date = Date(),
                 defaultDiscount = disc,
-                minimumInvoiceAmount = BigDecimal.valueOf(random.nextInt(100) * 1000.0),
-                totalInvestment = BigDecimal.valueOf(900000000.0),
-                maximumInvoiceAmount = BigDecimal.valueOf(750000.0)
+                minimumInvoiceAmount = random.nextInt(100) * 1000.0,
+                totalInvestment = 900000000.0,
+                maximumInvoiceAmount = 750000.0
         )
         val params: MutableMap<String, String> = mutableMapOf()
         params["issuedBy"] = "me"
