+++ bfn-atlas2/workflows/src/main/kotlin/com/bfn/flows/anchor/AnchorMakePaymentFlow.kt
@@ -12,7 +12,6 @@ import net.corda.core.contracts.StateAndRef
 import net.corda.core.flows.*
 import net.corda.core.transactions.TransactionBuilder
 import org.slf4j.LoggerFactory
-import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.ZoneId
 import java.util.*
@@ -85,7 +84,7 @@ class AnchorMakePaymentFlow(private val acceptedOffer: InvoiceOfferState) : Flow
             return null
         }
         val mDisc = disc / 100
-        val offerDouble = invoice.totalAmount * BigDecimal.valueOf(1.0 - mDisc)
+        val offerDouble = invoice.totalAmount * (1.0 - mDisc)
         return InvoiceOfferState(
                 invoiceId = invoice.invoiceId,
                 invoiceNumber = invoice.invoiceNumber,
